requirements.txt
Pillow==11.3.0
streamlit==1.42.2
torch==2.5.1+cu121
torchvision==0.20.1+cu121
